refactor(plugin): move wireRpcServerCleanup out of the plugin entry
OpenCode's plugin loader treats every named export from a plugin entry
point as a separate plugin factory and invokes it with `undefined`
input. The exported `wireRpcServerCleanup` was being called that way,
which would have crashed on the first reference to its callback if the
loader ever swept it up.

Move the helper to `src/lib/rpc-cleanup.ts` (matching the existing
`src/lib/` style alongside `ansi.ts`, `errno.ts`, `kill-tree.ts`,
`normalize-tool-arg-schemas.ts`). The plugin entry now imports it
internally and re-exports nothing beyond `default`.

Same lesson Systematic learned in v2.5.0 (PR #309) and again in
v2.12.1/v2.12.2 (PR #355): plugin entries must expose `export default`
only. Helpers belong in `src/lib/`.

No behavior change — the helper's identity-once semantics and its single
caller are byte-identical. Sets the stage for the Node-loadable build
(next commit) and the CI export-shape gate (subsequent commit).
